--- anova/data.csv
@@ -1,121 +1,0 @@
-y,x
-8,1
-44,2
-61,3
-149,4
-226,5
-263,6
-289,7
-452,8
-440,9
-699,10
-748,11
-896,12
-1039,13
-1214,14
-1377,15
-1580,16
-1763,17
-1994,18
-2197,19
-2456,20
-2679,21
-2967,22
-3208,23
-3526,24
-3785,25
-4133,26
-4410,27
-4788,28
-5082,29
-5491,30
-5803,31
-6243,32
-6571,33
-7043,34
-7387,35
-7891,36
-8251,37
-8788,38
-9162,39
-9732,40
-10121,41
-10725,42
-11128,43
-11767,44
-12183,45
-12856,46
-13285,47
-13994,48
-14435,49
-15180,50
-15633,51
-16414,52
-16879,53
-17696,54
-18173,55
-19027,56
-19514,57
-20406,58
-20903,59
-21833,60
-22340,61
-23308,62
-23824,63
-24832,64
-25357,65
-26403,66
-26937,67
-28023,68
-28564,69
-29692,70
-30240,71
-31408,72
-31963,73
-33173,74
-33734,75
-34986,76
-35553,77
-36847,78
-37420,79
-38757,80
-39334,81
-40715,82
-41296,83
-42721,84
-43306,85
-44775,86
-45364,87
-46877,88
-47469,89
-49028,90
-49623,91
-51227,92
-51824,93
-53474,94
-54072,95
-55769,96
-56369,97
-58113,98
-58713,99
-60505,100
-61105,101
-62945,102
-63545,103
-65434,104
-66032,105
-67970,106
-68567,107
-70555,108
-71150,109
-73188,110
-73781,111
-75869,112
-76460,113
-78599,114
-79186,115
-81377,116
-81960,117
-84203,118
-84782,119
-87077,120
